Made shared components, added page navigation, applied shared CSS to all relevant pages
