+++ frontend/src/pages/dashboard.js
@@ -1,11 +1,11 @@
 import "../styles/default.css";
 import { useState } from "react";
 
-// Import componets that will make up the profile page
+// Import componets that will make up the dashboard page
 import Header from "../components/header";
 import Sidebar from "../components/sidebar";
 
-function Profile() {
+function Dashboard() {
 	const [sidebarOpen, setSidebarOpen] = useState(true);
 
 	return (
@@ -18,16 +18,20 @@ function Profile() {
 				<Header
 					sidebarOpen={sidebarOpen}
 					onMenuToggle={() => setSidebarOpen(true)}
+					activeTab="Dashboard"
 				/>
 
-				{/* The title and brief description of the profile page  */}
+				{/* The title and brief description of the dashboard page  */}
 				<div className="hero-section">
-					<h2>Account Settings</h2>
-					<p>Manage your personal information</p>
+					<h2>Dashboard</h2>
+					<p>
+						View equipment usage, generate monthly reports, and add equipment to
+						the database
+					</p>
 				</div>
 			</div>
 		</div>
 	);
 }
 
-export default Profile;
+export default Dashboard;
