+++ frontend/src/pages/usermanagement.js
@@ -1,11 +1,11 @@
 import "../styles/default.css";
 import { useState } from "react";
 
-// Import componets that will make up the profile page
+// Import componets that will make up the user management page
 import Header from "../components/header";
 import Sidebar from "../components/sidebar";
 
-function Profile() {
+function UserManagement() {
 	const [sidebarOpen, setSidebarOpen] = useState(true);
 
 	return (
@@ -18,16 +18,17 @@ function Profile() {
 				<Header
 					sidebarOpen={sidebarOpen}
 					onMenuToggle={() => setSidebarOpen(true)}
+					activeTab="User Management"
 				/>
 
-				{/* The title and brief description of the profile page  */}
+				{/* The title and brief description of the user management page  */}
 				<div className="hero-section">
-					<h2>Account Settings</h2>
-					<p>Manage your personal information</p>
+					<h2>User Management</h2>
+					<p>Monitor user accounts and permissions</p>
 				</div>
 			</div>
 		</div>
 	);
 }
 
-export default Profile;
+export default UserManagement;
