+++ frontend/src/styles/home.css
@@ -1,227 +1,3 @@
-* {
-	box-sizing: border-box;
-	margin: 0;
-	padding: 0;
-}
-
-html,
-body,
-#root {
-	height: 100%;
-	width: 100%;
-	font-family: Arial, sans-serif;
-}
-
-.home-container {
-	display: flex;
-	height: 100%;
-	width: 100%;
-}
-
-/* Sidebar */
-.sidebar {
-	width: 280px;
-	background-color: #e3e3e3;
-	padding: 20px;
-	height: 100%;
-	overflow-y: auto;
-	flex-shrink: 0;
-}
-
-/* No visible scrollbars for modern design */
-.sidebar::-webkit-scrollbar {
-	display: none;
-}
-
-.sidebar {
-	-ms-overflow-style: none;
-	scrollbar-width: none;
-}
-
-/* Placeholder class for the sidebar back/close button */
-.back-button {
-	background: none;
-}
-
-.search-box {
-	position: relative;
-	margin-bottom: 30px;
-}
-
-.search-box input {
-	width: 100%;
-	padding: 10px 35px 10px 10px;
-	border: 1px solid #ccc;
-	border-radius: 4px;
-	font-size: 14px;
-}
-
-.sidebar h3 {
-	font-size: 16px;
-	margin: 20px 0 10px 0;
-	color: #333;
-}
-
-.farm-list {
-	display: flex;
-	flex-direction: column;
-	gap: 8px;
-	margin-bottom: 20px;
-}
-
-/* All of the quick-selection farm filters */
-.farm-item {
-	padding: 10px 15px;
-	background-color: white;
-	border: none;
-	border-radius: 4px;
-	text-align: left;
-	cursor: pointer;
-	font-size: 14px;
-	transition: background-color 0.2s;
-}
-
-.farm-item:hover {
-	background-color: #f0f0f0;
-}
-
-.farm-item.active {
-	background-color: #003d82;
-	color: white;
-	font-weight: 600;
-}
-
-/* Sidebar filtering styles */
-.filter-group {
-	margin-bottom: 15px;
-}
-
-.filter-group label {
-	display: block;
-	font-size: 13px;
-	margin-bottom: 5px;
-	color: #666;
-}
-
-.filter-group select {
-	width: 100%;
-	padding: 8px;
-	border: 1px solid #ccc;
-	border-radius: 4px;
-	background-color: white;
-	font-size: 14px;
-}
-
-/* Main */
-.main {
-	flex: 1;
-	display: flex;
-	flex-direction: column;
-	height: 100%;
-	overflow: hidden;
-}
-
-/* Header */
-.header {
-	background-color: white;
-	border-bottom: 1px solid #e0e0e0;
-}
-
-/* Defines a section for the sidebar toggle, title,
-notification and profile buttons */
-.header-top {
-	display: flex;
-	align-items: center;
-	justify-content: center;
-	padding: 15px 20px;
-	gap: 15px;
-	position: relative;
-}
-
-/* Opens the sidebar */
-/* Placeholder class for the sidebar open button */
-.menu-toggle {
-	position: absolute;
-	left: 20px;
-	background: none;
-}
-
-.header-top h1 {
-	font-size: 24px;
-	font-weight: 600;
-	text-align: center;
-}
-
-/* Styling for the section that the notification 
-and profile buttons are in */
-.header-right {
-	position: absolute;
-	right: 20px;
-	display: flex;
-	align-items: center;
-	gap: 15px;
-}
-
-/* Placeholder class for these icons */
-.notification-icon,
-.profile-icon {
-	width: 40px;
-	height: 40px;
-	background-color: #aaa;
-	border-radius: 50%;
-	display: flex;
-	justify-content: center;
-	align-items: center;
-	cursor: pointer;
-}
-
-/* Navigation items in the header */
-.nav-tabs {
-	display: flex;
-	justify-content: center;
-	gap: 10px;
-	padding: 10px 20px;
-	background-color: white;
-}
-
-.nav-tab {
-	padding: 10px 20px;
-	background-color: #f5f5f5;
-	border: 1px solid #ddd;
-	border-radius: 4px;
-	font-size: 14px;
-	cursor: pointer;
-	transition: all 0.2s;
-	color: #333;
-}
-
-.nav-tab:hover {
-	background-color: #e3e3e3;
-}
-
-.nav-tab.active {
-	background-color: #082e58;
-	color: white;
-	border-color: #082e58;
-}
-
-/* The title and brief description within each page */
-.hero-section {
-	background-color: #082e58;
-	color: white;
-	padding: 30px 25px;
-}
-
-.hero-section h2 {
-	font-size: 32px;
-	margin-bottom: 8px;
-}
-
-.hero-section p {
-	font-size: 16px;
-	opacity: 0.9;
-}
-
 /* Scrollable content within blue area 
 This will hold the equipment card items */
 .content {
